Message transformation improvements,
